% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/8_Functions_test_package_4Methods_random_XYsites.R
\name{check.inputs}
\alias{check.inputs}
\title{Checks input raster layers}
\usage{
check.inputs(occ)
}
\arguments{
\item{occ}{Run this function requires loading in the occupancy raster stack}
}
\description{
This function runs a number of checks before running the power analysis simulations to make sure the raster layers have the same dimensions
}
