% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/8_Functions_test_package_4Methods_random_XYsites.R
\name{fire.point.model2}
\alias{fire.point.model2}
\title{Stochastic disturbance function}
\usage{
fire.point.model2(time.hist, jj, Pr.dist, n.sites)
}
\arguments{
\item{time.hist}{An array specifying the time since a disturbance at each monitoring site prior to simulations}

\item{jj}{The year of simulation. Any value from 1 to Tmax}

\item{Pr.dist}{A vector specifying the probability of a cell being disturbed given the time since the last disturbance.}

\item{n.sites}{The number of sites to be monitored each simulation}
}
\description{
This function models the incidence of a stochastic disturbance at monitoring sites given the probability of a cell being disturbed as a function of time since disturbance.
At present, separate hazard functions are defined for three different vegetation classes. It returns a vector specifying whether each site burns (1) or not (0) in a given year jj
}
