% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/8_Functions_test_package_4Methods_random_XYsites.R
\name{plot.power}
\alias{plot.power}
\title{Plot results of power analysis}
\usage{
plot.power(pwr, n.species, nsims, effect.size, n.park, species.list, n.cores)
}
\arguments{
\item{pwr}{An array containing the results of the power analysis}

\item{n.species}{The number of species}

\item{nsims}{The number of simulations}

\item{effect.size}{The proportional decline in occupancy}

\item{n.park}{The number of parks in which to restimate power}

\item{species.list}{A vector specifying the name of each species analysed}

\item{save.wd}{A directory to save results and figures}
}
\description{
This function plots results of the power analysis. It first unpacks the list generated by the foreach function and estimates power given the number of simulations 
and the number of cores used. Power is plotted for each species on a single figure. Users can then scroll through separate plots for power at a park level if it park.power = TRUE
}
\examples{
plot.results()
}
\keyword{cats}
