% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/8_Functions_test_package_4Methods_random_XYsites.R
\name{run.power}
\alias{run.power}
\title{Run power analysis}
\usage{
run.power(effect.size, nsims, alpha, Tmax, s.years, trend, sites,
  model.disturbance, disturbance.type, species.list, dist.occ.time,
  dist.occ.freq, park.level, xy.sites, R, two.tailed, plot.sites, n.sites,
  n.species, n.park, all.loaded.sites, load.sites, new.site.selection, occ.time,
  det.method1.time, det.method2.time, det.method3.time, det.method4.time,
  n.method, occ, det.method1, det.method2, det.method3, det.method4, stratify,
  remote, parks, Pr.dist, disturbance)
}
\arguments{
\item{effect.size}{An integer specifying the proportional reduction in occupancy between the start and end of the monitoring program}

\item{nsims}{An integer specifying the number of simulations}

\item{alpha}{The Type I error rate. Must be either 0.1, 0.05 or 0.01}

\item{Tmax}{An integer specifying the length of the monitoring program}

\item{s.years}{A vector specifying the years in which monitoring occurs. Note, the final year or monitoring must be equal to Tmax}

\item{sites}{An array containing the XY coordinates of pre-specified monitoring sites}

\item{species.list}{An array containing the name of each species in the first column, and a zero or one describing how many methods are used to detect the species}

\item{park.level}{Set to TRUE if power is to be estimated within sub-level management units}

\item{xy.sites}{The XY coordinates of monitoring sites}

\item{two.tailed}{Set to TRUE is conducting a two-tailed significance test, FALSE if conducting a one-tailed test}

\item{plot.sites}{Set to TRUE to plot the location of sites at the start of each simulation, FALSE otherwise}

\item{n.sites}{The number of sites to be monitored.}

\item{n.species}{The number of species for which monitoring is simulated. Is equal to the number of layers in the occ raster stack}

\item{n.park}{The number of sub-level management units in which power is estimated}

\item{all.loaded.sites}{Set to TRUE if all of the prespecified sites are to be monitored, FALSE if fewer or more sites are to be monitored}

\item{occ.time}{An array containing the occupancy value for each species at monitoring sites over time}

\item{det.method1.time}{An array containing the detectability values of method 1 for each species at monitoring sites over time}

\item{det.method2.time}{An array containing the detectability values of method 2 for each species at monitoring sites over time}

\item{det.method3.time}{An array containing the detectability values of method 3 for each species at monitoring sites over time}

\item{decline}{Specifies the pattern of decline across space. Set to 'random' to model a constant 'blanket' decline across space}

\item{fire.hist}{Delete}

\item{model.fire}{Set to TRUE if the incidence of fire is modelled at sites, FALSE to model a static landscape over time}

\item{fire.fr}{Delete}

\item{combined.effect}{An array that records the reduction in occupancy across sites due to the effect size and fire}

\item{randomise.sites}{Set to TRUE if sites are to randomly selected for monitoring, FALSE if pre-specified sites are surveyed}

\item{occ.new}{An array containing the occupancy value for each species at monitoring sites over time}

\item{det.method1.new}{An array containing the detectability values of method 1 for each species at monitoring sites over time}

\item{det.method2.new}{An array containing the detectability values of method 2 for each species at monitoring sites over time}

\item{det.method3.new}{An array containing the detectability values of method 3 for each species at monitoring sites over time}

\item{det1.method1}{An array that records simulated detection histories at sites using method 1}

\item{det1.method2}{An array that records simulated detection histories at sites using method 2}

\item{det1.method3}{An array that records simulated detection histories at sites using method 3}

\item{fire.freq}{Delete!}

\item{n.method1}{The number of repeat visits to sites using method 1}

\item{n.method2}{The number of repeat visits to sites using method 2}

\item{n.method3}{The number of repeat visits to sites using method 3}
}
\description{
This function runs the main power analysis. It starts by creating empty vectors to record the number of times a significant trend is detected in the simulated detection histories
It then simulates fire at sites if model.fire = TRUE, and models a decline in occupancy over time depending on the specified effect size. 
Detection histories are simulated at each site given the occupancy status at each site, the number of survey methods used to detect a species and the detection probability of each relevent method.
Simulated detection histories are loaded into the package unmarked, and a trend in occupancy is modelled over time. The proportion of times a significant trend is detected is recorded.
This process is repeated for each effect size.
}
\examples{
run.power()
}
